Appendix B - Source code/modules/FileCreator/requirements.txt
azure-iot-device~=2.7.0
numpy~=1.21.5
azure-storage-blob~=12.11.0